+++ gneiss-mqtt/src/features/gneiss_tokio/mod.rs
@@ -8,6 +8,8 @@ Functionality for using [`tokio`](https://crates.io/crates/tokio) as an MQTT cli
 runtime implementation.
  */
 
+#[cfg(all(feature = "testing", test))]
+mod longtests;
 
 use crate::client::*;
 use crate::config::*;
@@ -604,9 +606,8 @@ impl Drop for ClientEventWaiter {
     }
 }
 
-#[cfg(test)]
+#[cfg(all(test, feature = "testing"))]
 pub(crate) mod testing {
-    use std::collections::VecDeque;
     use assert_matches::assert_matches;
     use std::future::Future;
     use std::pin::Pin;
@@ -614,16 +615,12 @@ pub(crate) mod testing {
     use std::time::Duration;
     use crate::client::waiter::*;
     use crate::config::*;
-    use crate::error::{MqttError, MqttResult};
+    use crate::error::*;
     use crate::mqtt::*;
     use super::*;
-    #[cfg(feature = "testenv")]
     use crate::testing::integration::*;
-    use crate::testing::mock_server::*;
-    use crate::testing::protocol::{BrokerTestContext, create_default_packet_handlers};
 
-    #[cfg(feature = "testenv")]
-    fn create_client_builder_internal(connect_options: ConnectOptions, tls_config: TlsUsage, ws_config: WebsocketUsage, proxy_config: ProxyUsage, tls_endpoint: TlsUsage, ws_endpoint: WebsocketUsage) -> GenericClientBuilder {
+    fn create_client_builder_internal(connect_options: ConnectOptions, _tls_usage: TlsUsage, ws_config: WebsocketUsage, proxy_config: ProxyUsage, tls_endpoint: TlsUsage, ws_endpoint: WebsocketUsage) -> GenericClientBuilder {
         let client_config = Mqtt5ClientOptionsBuilder::new()
             .with_connect_timeout(Duration::from_secs(5))
             .with_offline_queue_policy(OfflineQueuePolicy::PreserveAll)
@@ -636,7 +633,8 @@ pub(crate) mod testing {
         builder.with_connect_options(connect_options);
         builder.with_client_options(client_config);
 
-        if tls_config != TlsUsage::None {
+        #[cfg(feature = "rustls")]
+        if _tls_usage != TlsUsage::None {
             let mut tls_options_builder = TlsOptionsBuilder::new();
             tls_options_builder.with_verify_peer(false);
             tls_options_builder.with_root_ca_from_path(&get_ca_path()).unwrap();
@@ -659,7 +657,6 @@ pub(crate) mod testing {
         builder
     }
 
-    #[cfg(feature = "testenv")]
     fn create_good_client_builder(tls: TlsUsage, ws: WebsocketUsage, proxy: ProxyUsage) -> GenericClientBuilder {
         let connect_options = ConnectOptionsBuilder::new()
             .with_rejoin_session_policy(RejoinSessionPolicy::PostSuccess)
@@ -672,7 +669,6 @@ pub(crate) mod testing {
     type AsyncTestFactoryReturnType = Pin<Box<dyn Future<Output = MqttResult<()>> + Send>>;
     type AsyncTestFactory = Box<dyn Fn(GenericClientBuilder) -> AsyncTestFactoryReturnType + Send + Sync>;
 
-    #[cfg(feature = "testenv")]
     fn do_good_client_test(tls: TlsUsage, ws: WebsocketUsage, proxy: ProxyUsage, test_factory: AsyncTestFactory) {
         let runtime = tokio::runtime::Runtime::new().unwrap();
         let test_future = (*test_factory)(create_good_client_builder(tls, ws, proxy));
@@ -680,14 +676,13 @@ pub(crate) mod testing {
         runtime.block_on(test_future).unwrap();
     }
 
-    fn do_builder_test(test_factory: AsyncTestFactory, builder: GenericClientBuilder) {
+    pub(crate) fn do_builder_test(test_factory: AsyncTestFactory, builder: GenericClientBuilder) {
         let runtime = tokio::runtime::Runtime::new().unwrap();
         let test_future = (*test_factory)(builder);
 
         runtime.block_on(test_future).unwrap();
     }
 
-    #[cfg(feature = "testenv")]
     async fn start_client(client: &Arc<Mqtt5Client>) -> MqttResult<()> {
         let mut connection_attempt_waiter = ClientEventWaiter::new_single(client.clone(), ClientEventType::ConnectionAttempt);
         let mut connection_success_waiter = ClientEventWaiter::new_single(client.clone(), ClientEventType::ConnectionSuccess);
@@ -708,7 +703,6 @@ pub(crate) mod testing {
         Ok(())
     }
 
-    #[cfg(feature = "testenv")]
     async fn stop_client(client: &Arc<Mqtt5Client>) -> MqttResult<()> {
         let mut disconnection_waiter = ClientEventWaiter::new_single(client.clone(), ClientEventType::Disconnection);
         let mut stopped_waiter = ClientEventWaiter::new_single(client.clone(), ClientEventType::Stopped);
@@ -730,7 +724,6 @@ pub(crate) mod testing {
         Ok(())
     }
 
-    #[cfg(feature = "testenv")]
     async fn tokio_connect_disconnect_test(builder: GenericClientBuilder) -> MqttResult<()> {
         let client = Arc::new(builder.build(&tokio::runtime::Handle::current()).unwrap());
 
@@ -740,7 +733,6 @@ pub(crate) mod testing {
         Ok(())
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
     fn client_connect_disconnect_direct_plaintext_no_proxy() {
         do_good_client_test(TlsUsage::None, WebsocketUsage::None, ProxyUsage::None, Box::new(|builder|{
@@ -748,15 +740,14 @@ pub(crate) mod testing {
         }));
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
+    #[cfg(feature = "rustls")]
     fn client_connect_disconnect_direct_rustls_no_proxy() {
         do_good_client_test(TlsUsage::Rustls, WebsocketUsage::None, ProxyUsage::None, Box::new(|builder|{
             Box::pin(tokio_connect_disconnect_test(builder))
         }));
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
     fn client_connect_disconnect_websocket_plaintext_no_proxy() {
         do_good_client_test(TlsUsage::None, WebsocketUsage::Tungstenite, ProxyUsage::None, Box::new(|builder|{
@@ -764,15 +755,14 @@ pub(crate) mod testing {
         }));
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
+    #[cfg(feature = "rustls")]
     fn client_connect_disconnect_websocket_rustls_no_proxy() {
         do_good_client_test(TlsUsage::Rustls, WebsocketUsage::Tungstenite, ProxyUsage::None, Box::new(|builder|{
             Box::pin(tokio_connect_disconnect_test(builder))
         }));
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
     fn client_connect_disconnect_direct_plaintext_with_proxy() {
         do_good_client_test(TlsUsage::None, WebsocketUsage::None, ProxyUsage::Plaintext, Box::new(|builder|{
@@ -780,15 +770,14 @@ pub(crate) mod testing {
         }));
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
+    #[cfg(feature = "rustls")]
     fn client_connect_disconnect_direct_rustls_with_proxy() {
         do_good_client_test(TlsUsage::Rustls, WebsocketUsage::None, ProxyUsage::Plaintext, Box::new(|builder|{
             Box::pin(tokio_connect_disconnect_test(builder))
         }));
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
     fn client_connect_disconnect_websocket_plaintext_with_proxy() {
         do_good_client_test(TlsUsage::None, WebsocketUsage::Tungstenite, ProxyUsage::Plaintext, Box::new(|builder|{
@@ -796,15 +785,14 @@ pub(crate) mod testing {
         }));
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
+    #[cfg(feature = "rustls")]
     fn client_connect_disconnect_websocket_rustls_with_proxy() {
         do_good_client_test(TlsUsage::Rustls, WebsocketUsage::Tungstenite, ProxyUsage::Plaintext, Box::new(|builder|{
             Box::pin(tokio_connect_disconnect_test(builder))
         }));
     }
 
-    #[cfg(feature = "testenv")]
     async fn tokio_subscribe_unsubscribe_test(builder: GenericClientBuilder) -> MqttResult<()> {
         let client = Arc::new(builder.build(&tokio::runtime::Handle::current()).unwrap());
         start_client(&client).await?;
@@ -836,7 +824,6 @@ pub(crate) mod testing {
         Ok(())
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
     fn client_subscribe_unsubscribe() {
         do_good_client_test(TlsUsage::None, WebsocketUsage::None, ProxyUsage::None, Box::new(|builder|{
@@ -844,7 +831,6 @@ pub(crate) mod testing {
         }));
     }
 
-    #[cfg(feature = "testenv")]
     fn verify_successful_publish_result(result: &PublishResponse, qos: QualityOfService) {
         match result {
             PublishResponse::Qos0 => {
@@ -864,7 +850,6 @@ pub(crate) mod testing {
         }
     }
 
-    #[cfg(feature = "testenv")]
     fn verify_publish_received(event: &PublishReceivedEvent, expected_topic: &str, expected_qos: QualityOfService, expected_payload: &[u8]) {
         let publish = &event.publish;
 
@@ -873,7 +858,6 @@ pub(crate) mod testing {
         assert_eq!(expected_payload, publish.payload.as_ref().unwrap().as_slice());
     }
 
-    #[cfg(feature = "testenv")]
     async fn tokio_subscribe_publish_test(builder: GenericClientBuilder, qos: QualityOfService) -> MqttResult<()> {
         let client = Arc::new(builder.build(&tokio::runtime::Handle::current()).unwrap());
         start_client(&client).await?;
@@ -913,7 +897,6 @@ pub(crate) mod testing {
         Ok(())
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
     fn client_subscribe_publish_qos0() {
         do_good_client_test(TlsUsage::None, WebsocketUsage::None, ProxyUsage::None, Box::new(|builder|{
@@ -921,7 +904,6 @@ pub(crate) mod testing {
         }));
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
     fn client_subscribe_publish_qos1() {
         do_good_client_test(TlsUsage::None, WebsocketUsage::None, ProxyUsage::None, Box::new(|builder|{
@@ -929,7 +911,6 @@ pub(crate) mod testing {
         }));
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
     fn client_subscribe_publish_qos2() {
         do_good_client_test(TlsUsage::None, WebsocketUsage::None, ProxyUsage::None, Box::new(|builder|{
@@ -937,7 +918,6 @@ pub(crate) mod testing {
         }));
     }
 
-    #[cfg(feature = "testenv")]
     // This primarily tests that the will configuration works.  Will functionality is mostly broker-side.
     async fn tokio_will_test(builder: GenericClientBuilder) -> MqttResult<()> {
         let client = Arc::new(builder.build(&tokio::runtime::Handle::current()).unwrap());
@@ -987,7 +967,6 @@ pub(crate) mod testing {
         Ok(())
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
     fn client_will_sent() {
         do_good_client_test(TlsUsage::None, WebsocketUsage::None, ProxyUsage::None, Box::new(|builder|{
@@ -995,7 +974,6 @@ pub(crate) mod testing {
         }));
     }
 
-    #[cfg(feature = "testenv")]
     async fn tokio_connect_disconnect_cycle_session_rejoin_test(builder: GenericClientBuilder) -> MqttResult<()> {
         let client = Arc::new(builder.build(&tokio::runtime::Handle::current()).unwrap());
         start_client(&client).await?;
@@ -1024,7 +1002,6 @@ pub(crate) mod testing {
         Ok(())
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
     fn connect_disconnect_cycle_session_rejoin() {
         do_good_client_test(TlsUsage::None, WebsocketUsage::None, ProxyUsage::None, Box::new(|builder|{
@@ -1032,7 +1009,6 @@ pub(crate) mod testing {
         }));
     }
 
-    #[cfg(feature = "testenv")]
     async fn connection_failure_test(builder : GenericClientBuilder) -> MqttResult<()> {
         let client = Arc::new(builder.build(&tokio::runtime::Handle::current()).unwrap());
         let mut connection_failure_waiter = ClientEventWaiter::new_single(client.clone(), ClientEventType::ConnectionFailure);
@@ -1045,7 +1021,6 @@ pub(crate) mod testing {
         Ok(())
     }
 
-    #[cfg(feature = "testenv")]
     fn create_mismatch_builder(tls_config: TlsUsage, ws_config: WebsocketUsage, tls_endpoint: TlsUsage, ws_endpoint: WebsocketUsage) -> GenericClientBuilder {
         assert!(tls_config != tls_endpoint || ws_config != ws_endpoint);
 
@@ -1054,8 +1029,8 @@ pub(crate) mod testing {
         create_client_builder_internal(connect_options, tls_config, ws_config, ProxyUsage::None, tls_endpoint, ws_endpoint)
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
+    #[cfg(feature = "rustls")]
     fn connection_failure_direct_tls_config_direct_plaintext_endpoint() {
         let builder = create_mismatch_builder(TlsUsage::Rustls, WebsocketUsage::None, TlsUsage::None, WebsocketUsage::None);
         do_builder_test(Box::new(|builder| {
@@ -1063,8 +1038,8 @@ pub(crate) mod testing {
         }), builder);
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
+    #[cfg(feature = "rustls")]
     fn connection_failure_direct_tls_config_websocket_plaintext_endpoint() {
         let builder = create_mismatch_builder(TlsUsage::Rustls, WebsocketUsage::None, TlsUsage::None, WebsocketUsage::Tungstenite);
         do_builder_test(Box::new(|builder| {
@@ -1072,8 +1047,8 @@ pub(crate) mod testing {
         }), builder);
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
+    #[cfg(feature = "rustls")]
     fn connection_failure_direct_tls_config_websocket_tls_endpoint() {
         let builder = create_mismatch_builder(TlsUsage::Rustls, WebsocketUsage::None, TlsUsage::Rustls, WebsocketUsage::Tungstenite);
         do_builder_test(Box::new(|builder| {
@@ -1081,8 +1056,8 @@ pub(crate) mod testing {
         }), builder);
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
+    #[cfg(feature = "rustls")]
     fn connection_failure_direct_plaintext_config_direct_tls_endpoint() {
         let builder = create_mismatch_builder(TlsUsage::None, WebsocketUsage::None, TlsUsage::Rustls, WebsocketUsage::None);
         do_builder_test(Box::new(|builder| {
@@ -1090,7 +1065,6 @@ pub(crate) mod testing {
         }), builder);
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
     fn connection_failure_direct_plaintext_config_websocket_plaintext_endpoint() {
         let builder = create_mismatch_builder(TlsUsage::None, WebsocketUsage::None, TlsUsage::None, WebsocketUsage::Tungstenite);
@@ -1099,8 +1073,8 @@ pub(crate) mod testing {
         }), builder);
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
+    #[cfg(feature = "rustls")]
     fn connection_failure_direct_plaintext_config_websocket_tls_endpoint() {
         let builder = create_mismatch_builder(TlsUsage::None, WebsocketUsage::None, TlsUsage::Rustls, WebsocketUsage::Tungstenite);
         do_builder_test(Box::new(|builder| {
@@ -1108,8 +1082,8 @@ pub(crate) mod testing {
         }), builder);
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
+    #[cfg(feature = "rustls")]
     fn connection_failure_websocket_tls_config_direct_plaintext_endpoint() {
         let builder = create_mismatch_builder(TlsUsage::Rustls, WebsocketUsage::Tungstenite, TlsUsage::None, WebsocketUsage::None);
         do_builder_test(Box::new(|builder| {
@@ -1117,8 +1091,8 @@ pub(crate) mod testing {
         }), builder);
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
+    #[cfg(feature = "rustls")]
     fn connection_failure_websocket_tls_config_websocket_plaintext_endpoint() {
         let builder = create_mismatch_builder(TlsUsage::Rustls, WebsocketUsage::Tungstenite, TlsUsage::None, WebsocketUsage::Tungstenite);
         do_builder_test(Box::new(|builder| {
@@ -1126,8 +1100,8 @@ pub(crate) mod testing {
         }), builder);
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
+    #[cfg(feature = "rustls")]
     fn connection_failure_websocket_tls_config_direct_tls_endpoint() {
         let builder = create_mismatch_builder(TlsUsage::Rustls, WebsocketUsage::Tungstenite, TlsUsage::Rustls, WebsocketUsage::None);
         do_builder_test(Box::new(|builder| {
@@ -1135,7 +1109,6 @@ pub(crate) mod testing {
         }), builder);
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
     fn connection_failure_websocket_plaintext_config_direct_plaintext_endpoint() {
         let builder = create_mismatch_builder(TlsUsage::None, WebsocketUsage::Tungstenite, TlsUsage::None, WebsocketUsage::None);
@@ -1144,8 +1117,8 @@ pub(crate) mod testing {
         }), builder);
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
+    #[cfg(feature = "rustls")]
     fn connection_failure_websocket_plaintext_config_websocket_tls_endpoint() {
         let builder = create_mismatch_builder(TlsUsage::None, WebsocketUsage::Tungstenite, TlsUsage::Rustls, WebsocketUsage::Tungstenite);
         do_builder_test(Box::new(|builder| {
@@ -1153,8 +1126,8 @@ pub(crate) mod testing {
         }), builder);
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
+    #[cfg(feature = "rustls")]
     fn connection_failure_websocket_plaintext_config_direct_tls_endpoint() {
         let builder = create_mismatch_builder(TlsUsage::None, WebsocketUsage::Tungstenite, TlsUsage::Rustls, WebsocketUsage::None);
         do_builder_test(Box::new(|builder| {
@@ -1162,7 +1135,6 @@ pub(crate) mod testing {
         }), builder);
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
     fn connection_failure_invalid_endpoint() {
         let client_options = Mqtt5ClientOptionsBuilder::new()
@@ -1177,7 +1149,6 @@ pub(crate) mod testing {
         }), builder);
     }
 
-    #[cfg(feature = "testenv")]
     #[test]
     fn connection_failure_invalid_endpoint_http() {
         let builder = GenericClientBuilder::new("amazon.com", 443);
@@ -1186,264 +1157,5 @@ pub(crate) mod testing {
         }), builder);
     }
 
-    fn is_reconnect_related_event(event: &Arc<ClientEvent>) -> bool {
-        match **event {
-            ClientEvent::Stopped(_) | ClientEvent::PublishReceived(_) => {
-                false
-            }
-            _ => {
-                true
-            }
-        }
-    }
 
-    type ReconnectEventTestValidatorFn = Box<dyn Fn(&Vec<ClientEventRecord>) -> MqttResult<()> + Send + Sync>;
-
-    async fn simple_reconnect_test(builder : GenericClientBuilder, event_count: usize, event_checker: ReconnectEventTestValidatorFn) -> MqttResult<()> {
-        let client = Arc::new(builder.build(&tokio::runtime::Handle::current()).unwrap());
-
-        let wait_options = ClientEventWaiterOptions {
-            wait_type: ClientEventWaitType::Predicate(Box::new(|event|{ is_reconnect_related_event(event) }))
-        };
-
-        let mut reconnect_waiter = ClientEventWaiter::new(client.clone(), wait_options, event_count);
-
-        client.start(None)?;
-
-        let reconnect_events = reconnect_waiter.wait().await?;
-
-        client.stop(None)?;
-        client.close()?;
-
-        (event_checker)(&reconnect_events)
-    }
-
-    fn build_reconnect_test_options() -> ClientTestOptions {
-        let mut test_options = ClientTestOptions::default();
-
-        test_options.client_options_mutator_fn = Some(Box::new(|builder| {
-            builder.with_base_reconnect_period(Duration::from_millis(250));
-            builder.with_max_reconnect_period(Duration::from_millis(6000));
-            builder.with_reconnect_period_jitter(ExponentialBackoffJitterType::None);
-            builder.with_reconnect_stability_reset_period(Duration::from_millis(5000));
-        }));
-
-        test_options.packet_handler_set_factory_fn = Some(Box::new(|| {
-            let mut handlers = create_default_packet_handlers();
-
-            handlers.insert(PacketType::Connect, Box::new(crate::testing::protocol::handle_connect_with_failure_connack));
-
-            handlers
-        }));
-
-        test_options
-    }
-
-    fn validate_reconnect_failure_sequence(events: &Vec<ClientEventRecord>) -> MqttResult<()> {
-        let mut connection_failures: usize = 0;
-        let mut previous_failure_time : Option<Instant> = None;
-        let mut actual_delays : Vec<Duration> = Vec::new();
-
-        for (i, event_record) in events.iter().enumerate() {
-            if i % 2 == 0 {
-                assert_matches!(*event_record.event, ClientEvent::ConnectionAttempt(_));
-                if let Some(previous_timestamp) = &previous_failure_time {
-                    assert!(*previous_timestamp < event_record.timestamp);
-                    actual_delays.push(event_record.timestamp - *previous_timestamp);
-                }
-            } else {
-                assert_matches!(*event_record.event, ClientEvent::ConnectionFailure(_));
-                connection_failures += 1;
-                if let Some(old_failure_time) = previous_failure_time {
-                    assert!(old_failure_time < event_record.timestamp);
-                }
-                previous_failure_time = Some(event_record.timestamp);
-            }
-        }
-
-        assert_eq!(7, connection_failures);
-
-        let expected_delays : Vec<Duration> = vec!(250, 500, 1000, 2000, 4000, 6000).into_iter().map(|val| {Duration::from_millis(val)}).collect();
-        assert_eq!(expected_delays.len(), actual_delays.len());
-
-        let zipped_iter = expected_delays.iter().zip(actual_delays.iter());
-
-        for (_, (expected_delay, actual_delay)) in zipped_iter.enumerate() {
-            assert!(*actual_delay >= *expected_delay);
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn client_reconnect_with_backoff() {
-        let (builder, server) = build_mock_client_server(build_reconnect_test_options());
-
-        do_builder_test(Box::new(|builder| {
-            Box::pin(simple_reconnect_test(builder, 14, Box::new(|events|{validate_reconnect_failure_sequence(events)})))
-        }), builder);
-
-        server.close();
-    }
-
-    pub(crate) fn handle_connect_with_conditional_connack(packet: &Box<MqttPacket>, response_packets: &mut VecDeque<Box<MqttPacket>>, context: &mut BrokerTestContext) -> MqttResult<()> {
-        if let MqttPacket::Connect(_) = &**packet {
-            context.connect_count += 1;
-
-            if context.connect_count < 6 {
-                let response = Box::new(MqttPacket::Connack(ConnackPacket {
-                    reason_code: ConnectReasonCode::Banned,
-                    ..Default::default()
-                }));
-                response_packets.push_back(response);
-            } else {
-                let response = Box::new(MqttPacket::Connack(ConnackPacket {
-                    reason_code: ConnectReasonCode::Success,
-                    assigned_client_identifier: Some("client-id".to_string()),
-                    ..Default::default()
-                }));
-                response_packets.push_back(response);
-            }
-
-            return Ok(());
-        }
-
-        panic!("Invalid packet handler state")
-    }
-
-    pub(crate) fn handle_publish_with_disconnect(packet: &Box<MqttPacket>, response_packets: &mut VecDeque<Box<MqttPacket>>, _: &mut BrokerTestContext) -> MqttResult<()> {
-        if let MqttPacket::Publish(_) = &**packet {
-            let response = Box::new(MqttPacket::Disconnect(DisconnectPacket {
-                reason_code: DisconnectReasonCode::NotAuthorized,
-                ..Default::default()
-            }));
-            response_packets.push_back(response);
-
-            return Ok(());
-        }
-
-        panic!("Invalid packet handler state")
-    }
-
-    fn build_reconnect_reset_test_options() -> ClientTestOptions {
-        let mut test_options = ClientTestOptions::default();
-
-        test_options.client_options_mutator_fn = Some(Box::new(|builder| {
-            builder.with_base_reconnect_period(Duration::from_millis(500));
-            builder.with_max_reconnect_period(Duration::from_millis(6000));
-            builder.with_reconnect_period_jitter(ExponentialBackoffJitterType::None);
-            builder.with_reconnect_stability_reset_period(Duration::from_millis(3000));
-        }));
-
-        test_options.packet_handler_set_factory_fn = Some(Box::new(|| {
-            let mut handlers = create_default_packet_handlers();
-
-            handlers.insert(PacketType::Connect, Box::new(handle_connect_with_conditional_connack));
-            handlers.insert(PacketType::Publish, Box::new(handle_publish_with_disconnect));
-
-            handlers
-        }));
-
-        test_options
-    }
-
-    async fn reconnect_backoff_reset_test(builder : GenericClientBuilder, first_event_checker: ReconnectEventTestValidatorFn, second_event_checker_fn: ReconnectEventTestValidatorFn, connection_success_wait_millis: u64) -> MqttResult<()> {
-        let client = Arc::new(builder.build(&tokio::runtime::Handle::current()).unwrap());
-
-        let first_wait_options = ClientEventWaiterOptions {
-            wait_type: ClientEventWaitType::Predicate(Box::new(|event|{ is_reconnect_related_event(event) }))
-        };
-
-        let mut first_reconnect_waiter = ClientEventWaiter::new(client.clone(), first_wait_options, 12);
-        let mut success_waiter = ClientEventWaiter::new_single(client.clone(), ClientEventType::ConnectionSuccess);
-
-        client.start(None)?;
-
-        let first_reconnect_events = first_reconnect_waiter.wait().await?;
-        let _ = success_waiter.wait().await?;
-
-        let final_wait_options = ClientEventWaiterOptions {
-            wait_type: ClientEventWaitType::Predicate(Box::new(|event|{ is_reconnect_related_event(event) }))
-        };
-
-        let mut final_reconnect_waiter = ClientEventWaiter::new(client.clone(), final_wait_options, 3);
-
-        tokio::time::sleep(Duration::from_millis(connection_success_wait_millis)).await;
-
-        let publish = PublishPacket::builder("hello/world".to_string(), QualityOfService::AtLeastOnce).build();
-        let _ = client.publish(publish, None);
-
-        let final_reconnect_events = final_reconnect_waiter.wait().await?;
-
-        client.stop(None)?;
-        client.close()?;
-
-        (first_event_checker)(&first_reconnect_events)?;
-        (second_event_checker_fn)(&final_reconnect_events)
-    }
-
-    fn validate_reconnect_backoff_failure_sequence(events: &Vec<ClientEventRecord>) -> MqttResult<()> {
-        let mut connection_failures: usize = 0;
-
-        for (i, event_record) in events.iter().enumerate() {
-            if i % 2 == 0 {
-                assert_matches!(*event_record.event, ClientEvent::ConnectionAttempt(_));
-            } else if i < 11 {
-                assert_matches!(*event_record.event, ClientEvent::ConnectionFailure(_));
-                connection_failures += 1;
-            } else {
-                assert_matches!(*event_record.event, ClientEvent::ConnectionSuccess(_));
-            }
-        }
-
-        assert_eq!(5, connection_failures);
-        Ok(())
-    }
-
-    fn validate_reconnect_backoff_reset_sequence(events: &Vec<ClientEventRecord>, expected_reconnect_delay: Duration) -> MqttResult<()> {
-
-        let record1 = &events[0];
-        let record2 = &events[1];
-        let record3 = &events[2];
-
-        assert_matches!(*record1.event, ClientEvent::Disconnection(_));
-        assert_matches!(*record2.event, ClientEvent::ConnectionAttempt(_));
-        assert_matches!(*record3.event, ClientEvent::ConnectionSuccess(_));
-
-        let reconnect_delay = record2.timestamp - record1.timestamp;
-
-        assert!(reconnect_delay >= expected_reconnect_delay && reconnect_delay < 2 * expected_reconnect_delay);
-
-        Ok(())
-    }
-
-    #[test]
-    #[cfg(feature = "longtests")]
-    fn client_reconnect_with_backoff_and_backoff_reset() {
-        let (builder, server) = build_mock_client_server(build_reconnect_reset_test_options());
-
-        do_builder_test(Box::new(|builder| {
-            Box::pin(reconnect_backoff_reset_test(builder,
-                                                  Box::new(|events|{validate_reconnect_backoff_failure_sequence(events)}),
-                                                  Box::new(|events|{validate_reconnect_backoff_reset_sequence(events, Duration::from_millis(500))}),
-                                                  4000))
-        }), builder);
-
-        server.close();
-    }
-
-    #[test]
-    #[cfg(feature = "longtests")]
-    fn client_reconnect_with_backoff_and_no_backoff_reset() {
-        let (builder, server) = build_mock_client_server(build_reconnect_reset_test_options());
-
-        do_builder_test(Box::new(|builder| {
-            Box::pin(reconnect_backoff_reset_test(builder,
-                                                  Box::new(|events|{validate_reconnect_backoff_failure_sequence(events)}),
-                                                  Box::new(|events|{validate_reconnect_backoff_reset_sequence(events, Duration::from_millis(6000))}),
-                                                      500))
-        }), builder);
-
-        server.close();
-    }
 }
