+++ gneiss-mqtt/src/features/mod.rs
@@ -7,7 +7,7 @@
 Module that encompasses feature-specific logic (primarily TLS and Async runtime).
  */
 
-#[cfg(rustls)]
+#[cfg(feature = "rustls")]
 pub mod gneiss_rustls;
 
 pub mod gneiss_tokio;
