+++ gneiss-mqtt/src/testing/integration.rs
@@ -8,6 +8,8 @@ use std::env;
 #[derive(Debug, Eq, PartialEq, Copy, Clone)]
 pub(crate) enum TlsUsage {
     None,
+
+    #[cfg(feature = "rustls")]
     Rustls
 }
 
@@ -23,6 +25,7 @@ pub(crate) enum ProxyUsage {
     Plaintext
 }
 
+#[cfg(feature = "rustls")]
 pub(crate) fn get_ca_path() -> String {
     env::var("GNEISS_MQTT_TEST_BROKER_CA_PATH").unwrap()
 }
