+++ gneiss-mqtt/src/testing/mod.rs
@@ -3,8 +3,9 @@
  * SPDX-License-Identifier: Apache-2.0.
  */
 
-#[cfg(feature = "testenv")]
+#[cfg(feature = "testing")]
 pub(crate) mod integration;
 pub(crate) mod protocol;
+#[cfg(feature = "testing")]
 pub(crate) mod mock_server;
